Former-commit-id: d094db33ec2507f96305dfcfc5de502f1bbff3df

+++ epub_editor/src/components/mainpage/Directory.vue
@@ -1,46 +1,56 @@
 <template>
-  <v-treeview
-    v-model="tree"
-    :open="initiallyOpen"
-    :items="items"
-    activatable
-    item-key="name"
-    open-on-click
-    @input="openFile"
-  >
-    <template slot="label" slot-scope="{ item }" >
-      <div  @mouseover="mouseAction(item,true);" @mouseleave="mouseAction(item,false);">
-          <v-icon small style="padding: 0 5px;" v-if="!item.file">
-            {{ 'mdi-folder' }}
-          </v-icon>
-          <v-icon v-else small style="padding: 0 5px;" >
-            {{ files[item.file] }}
-          </v-icon>
-          <span @click="openFile(item); saveDialog=true">{{ item.name }}</span>
-        <span id="toc" v-if="deleteBtn[item.name]">
-          <v-btn
-          class="align-self-center rounded-sm"
-          icon
-          x-small
-         @click="deleteChapter(item)">
-            <v-icon>mdi-trash-can-outline</v-icon>
-          </v-btn>
-        </span>
-      </div>
-    </template>
+  <div>
+    <v-treeview
+      v-model="tree"
+      :open="initiallyOpen"
+      :items="items"
+      activatable
+      item-key="name"
+      open-on-click
+      @input="openFile"
+    >
+      <template slot="label" slot-scope="{ item }" >
+        <div  @mouseover="mouseAction(item,true);" @mouseleave="mouseAction(item,false);">
+            <v-icon small style="padding: 0 5px;" v-if="!item.file">
+              {{ 'mdi-folder' }}
+            </v-icon>
+            <v-icon v-else small style="padding: 0 5px;" >
+              {{ files[item.file] }}
+            </v-icon>
+            <span @click="openFile(item);">{{ item.name }}</span>
+          <span id="toc" v-if="deleteBtn[item.name]">
+            <v-btn
+            class="align-self-center rounded-sm"
+            icon
+            x-small
+          @click="deleteDialog=true; deleteitem=item">
+              <v-icon>mdi-trash-can-outline</v-icon>
+            </v-btn>
+          </span>
+        </div>
+      </template>
+    </v-treeview>
 
     <Confirm 
-     :dialog="saveDialog"
-     title="confirm.save-title"
-     content1="confirm.save-content"
-     confirm="confirm.save-confirm"
-     cancel="confirm.save-cancel"
-     @cancel="saveDialog = false"
-     @confirm="saveFile"
-    />
-
-  </v-treeview>
-
+      :dialog="saveDialog"
+      title="confirm.save-title"
+      content1="confirm.save-content"
+      confirm="confirm.save-confirm"
+      cancel="confirm.save-cancel"
+      @cancel="saveFile(false)"
+      @confirm="saveFile(true)"
+      />
+    <Confirm 
+      :dialog="deleteDialog"
+      title="deleteChapter.title"
+      content1="deleteChapter.content-1"
+      content2="deleteChapter.content-2"
+      confirm="deleteChapter.delete-btn"
+      cancel="deleteChapter.cancel"
+      @cancel="deleteChapter(false)"
+      @confirm="deleteChapter(true)"
+      />
+  </div>
 </template>
 
 <script>
@@ -49,7 +59,6 @@ import * as edit from '@/functions/edit.js';
 import Vue from 'vue'
 import Confirm from '@/components/mainpage/Confirm'
 
-const { dialog } = require('electron').remote;
 const fs = require('fs');
 
 export default {
@@ -60,10 +69,9 @@ export default {
   watch: {
     items: function () {
       this.initiallyOpen = ['EPUB', 'text']
-      for (let j = 0; j < this.items[0].children[4].children.length; j++) {
-        if (this.items[0].children[4].children[j].file === "xhtml") {
+      for(let j=0;j<this.items[0].children[4].children.length;j++){
+        if(this.items[0].children[4].children[j].file==="xhtml"){
           Vue.set(this.deleteBtn, this.items[0].children[4].children[j].name, false);
-          //this.deleteBtn[this.deleteBtn,this.items[0].children[4].children[j].name.toString]=false;
         }
       }
     },
@@ -95,41 +103,49 @@ export default {
       },
       tree: [],
       deleteBtn: new Object(),
-      saveDialog: false
+      saveDialog: false,
+      deleteDialog:false,
+      deleteitem:{},
+      value : ''
     };
   },
   methods: {
-    deleteChapter(item){
-      console.log("deleteChapter");
+    deleteChapter(res){
+      this.deleteDialog=false;
+      if(!res)
+        return;
+      let item=this.deleteitem;
       fs.unlinkSync(item.dirPath, (err)=>{
         console.log(err);
       });
       Vue.nextTick();
+
       //content 변경
       let content= fs.readFileSync(this.$store.state.ebookDirectory+ '/EPUB/content.opf').toString();
-      let start=content.indexOf('<item id="'+item.name);
-      let end =content.indexOf("/>",start);
+      let startText='<item id="'+item.name;
+      let endText="/>"
+      let start=content.indexOf(startText);
+      let end =content.indexOf(endText,start);
       content=content.slice(0,start)+content.slice(end+3);
 
       start=content.indexOf('<itemref idref="'+item.name);
       end=content.indexOf("/>",start);
       content=content.slice(0,start)+content.slice(end+3);
 
-      console.log("content : ",content);
-      fs.writeFile(this.$store.state.ebookDirectory+ '/EPUB/content.opf', content, (err) => {
-            if (err) {
-              console.log(err);
-            }
-          });
+      fs.writeFileSync(this.$store.state.ebookDirectory+ '/EPUB/content.opf', content, (err) => {
+        if (err) {
+          console.log(err);
+        }
+      });
 
       //toc 변경
       let temp = fs.readFileSync(this.$store.state.ebookDirectory+ '/EPUB/toc.ncx').toString();
 
-      let searchTextStart='chapter';
-      let searchTextEnd='.xhtml';
-      start=item.name.indexOf(searchTextStart);
-      end=item.name.indexOf(searchTextEnd);
-      let searchNumber=item.name.slice(start+searchTextStart.length,end);
+      startText='chapter';
+      endText='.xhtml';
+      start=item.name.indexOf(startText);
+      end=item.name.indexOf(endText,start);
+      let searchNumber=item.name.slice(start+startText.length,end);
 
       let point=0;
       let searchDirTextStart='<navPoint id="navPoint-';
@@ -145,7 +161,7 @@ export default {
           let t='</navPoint>';
           let last=temp.indexOf(t,dirEnd);
           temp=temp.slice(0,dirStart)+temp.slice(last+t.length);
-          fs.writeFile(this.$store.state.ebookDirectory+ '/EPUB/toc.ncx', temp, (err) => {
+          fs.writeFileSync(this.$store.state.ebookDirectory+ '/EPUB/toc.ncx', temp, (err) => {
             if (err) {
               console.log(err);
             }
@@ -154,8 +170,9 @@ export default {
         }
         point=dirEnd;
       }
-      eventBus.$emit('toc');
+      
       this.$store.state.selectedFileDirectory='';
+      eventBus.$emit('toc');
     },
     mouseAction(item,set){
       if(item.file=="xhtml"){
@@ -164,6 +181,7 @@ export default {
       }
     },
     openFile: function (val) { // 디렉토리에서 선택한 파일을 텍스트로 읽는 함수
+      this.value = val;
       if (val.children) {
         return  // 폴더면 그냥 return
       } else {  // 파일을 클릭한 것이면
@@ -171,40 +189,42 @@ export default {
           // 처음 여는게 아니거나 다른 파일을 열려고 하는 것이면, 변경여부 확인하고 저장여부를 물어본다.
           let original = fs.readFileSync(this.$store.state.selectedFileDirectory).toString();  // original: 원래 작성중이던 파일의 원본
           original = original.slice(original.indexOf("<body"), original.indexOf("</body>") + 7);
-          if (original !== this.$store.state.editingText) {  // 원본과 수정 중 파일이 다르다면
-            // 저장 할 것인지 물어보고 저장 / 취소
-            
+          if (!this.compareFile(original, this.$store.state.editingText)) {
             this.saveDialog = true;
-            console.log('바뀌어서 다이얼로그 띄우고싶은데');
-            const options = {
-              type: 'question',
-              message: this.$t('confirm.save-title'),
-              buttons: [this.GET_CONFIRM_BUTTON, this.GET_CANCEL_BUTTON],
-            };
-            const result = dialog.showMessageBoxSync(options);
-            if (result === 0) {
-              const updatedText = this.$store.state.editingHTMLText + this.$store.state.editingText + '</html>';
-              fs.writeFileSync(this.$store.state.selectedFileDirectory, updatedText);
-              this.$store.dispatch('setAlertMessage', 'success.save-ebook');
-            }
-          } 
-        }
-        if (this.$store.state.selectedFileDirectory !== val.dirPath) {  // 자기 자신을 다시 클릭한 것이 아니라면 새 파일을 가져온다.
-          if (val.dirPath.slice(-5) === 'xhtml') {    // xhtml 파일이라면 불러온다.
-            const temp = fs.readFileSync(val.dirPath).toString();
-            this.$store.dispatch('setSelectedFileDirectory', val.dirPath);
-            eventBus.$emit('loadData', temp);
-          } else {    // xhtml 파일이 아니라면 alert 를 띄운다. 
-            this.$store.dispatch('setAlertMessage', 'error.select-text');
+          } else {
+            this.moveFile(val);
           }
         }
+        else {
+          this.moveFile(val);
+          }
+
         edit.reset();
       }
     },
-    saveFile: function () {
-      const updatedText = this.$store.state.editingHTMLText + this.$store.state.editingText + '</html>';
-      fs.writeFileSync(this.$store.state.selectedFileDirectory, updatedText);
-      this.$store.dispatch('setAlertMessage', 'success.save-ebook');
+    saveFile: function (flag) { // 저장
+      if (flag) {
+        const updatedText = this.$store.state.editingHTMLText + this.$store.state.editingText + '</html>';
+        fs.writeFileSync(this.$store.state.selectedFileDirectory, updatedText);
+        this.$store.dispatch('setAlertMessage', 'success.save-ebook');
+      } 
+      this.saveDialog = false;
+      this.moveFile(this.value);
+    },
+    moveFile: function (val) { // 파일 이동
+      if (this.$store.state.selectedFileDirectory !== val.dirPath) {  // 자기 자신을 다시 클릭한 것이 아니라면 새 파일을 가져온다.
+        if (val.dirPath.slice(-5) === 'xhtml') {    // xhtml 파일이라면 불러온다.
+          const temp = fs.readFileSync(val.dirPath).toString();
+          this.$store.dispatch('setSelectedFileDirectory', val.dirPath);
+          eventBus.$emit('loadData', temp);
+        } else {    // xhtml 파일이 아니라면 alert 를 띄운다. 
+          this.$store.dispatch('setAlertMessage', 'error.select-text');
+        }
+      }
+    },
+    compareFile: function (pre, post) { // 비교
+      if (pre === post) return true;
+      return false;
     }
   },
 }
